Move BaseModel from sarc.config to sarc.model
