+++ sarc/client/gpumetrics.py
@@ -7,7 +7,8 @@ from typing import Dict, List
 from pydantic import validator
 from pydantic_mongo import AbstractRepository, ObjectIdField
 
-from sarc.config import MTL, UTC, BaseModel, config, scraping_mode_required
+from sarc.config import MTL, UTC, config, scraping_mode_required
+from sarc.model import BaseModel
 
 logger = logging.getLogger(__name__)
